issue-25 - create new module to keep the balances and delete the erc20s;
