+++ docs/Hydra/RewardPool/modules/Vesting.md
@@ -253,6 +253,29 @@ function getMaxRSI() external pure returns (uint256 nominator)
 |---|---|---|
 | nominator | uint256 | undefined |
 
+### getRPSValues
+
+```solidity
+function getRPSValues(address validator, uint256 currentEpochId) external view returns (struct RPS[])
+```
+
+
+
+
+
+#### Parameters
+
+| Name | Type | Description |
+|---|---|---|
+| validator | address | undefined |
+| currentEpochId | uint256 | undefined |
+
+#### Returns
+
+| Name | Type | Description |
+|---|---|---|
+| _0 | RPS[] | undefined |
+
 ### getRSI
 
 ```solidity
@@ -376,6 +399,29 @@ function isActivePosition(address staker) external view returns (bool)
 | Name | Type | Description |
 |---|---|---|
 | staker | address | undefined |
+
+#### Returns
+
+| Name | Type | Description |
+|---|---|---|
+| _0 | bool | undefined |
+
+### isMaturingDelegatePosition
+
+```solidity
+function isMaturingDelegatePosition(address validator, address delegator) external view returns (bool)
+```
+
+
+
+
+
+#### Parameters
+
+| Name | Type | Description |
+|---|---|---|
+| validator | address | undefined |
+| delegator | address | undefined |
 
 #### Returns
