+++ docs/Hydra/ValidatorSet/modules/Delegation/IDelegation.md
@@ -45,10 +45,10 @@ Undelegates amount from validator. Apply penalty in case vesting is not finished
 | validator | address | Validator to undelegate from |
 | amount | uint256 | Amount to be undelegated |
 
-### delegate
+### delegateToValidator
 
 ```solidity
-function delegate(address validator, bool restake) external payable
+function delegateToValidator(address validator) external payable
 ```
 
 Delegates sent amount to validator. Claims rewards beforehand.
@@ -60,58 +60,11 @@ Delegates sent amount to validator. Claims rewards beforehand.
 | Name | Type | Description |
 |---|---|---|
 | validator | address | Validator to delegate to |
-| restake | bool | Whether to redelegate the claimed rewards |
 
-### delegationOf
-
-```solidity
-function delegationOf(address validator, address delegator) external view returns (uint256)
-```
-
-Gets amount delegated by delegator to validator.
-
-
-
-#### Parameters
-
-| Name | Type | Description |
-|---|---|---|
-| validator | address | Address of validator |
-| delegator | address | Address of delegator |
-
-#### Returns
-
-| Name | Type | Description |
-|---|---|---|
-| _0 | uint256 | Amount delegated (in MATIC wei) |
-
-### getDelegatorReward
+### openVestedDelegatePosition
 
 ```solidity
-function getDelegatorReward(address validator, address delegator) external view returns (uint256)
-```
-
-Gets delegators&#39;s unclaimed rewards with validator.
-
-
-
-#### Parameters
-
-| Name | Type | Description |
-|---|---|---|
-| validator | address | Address of validator |
-| delegator | address | Address of delegator |
-
-#### Returns
-
-| Name | Type | Description |
-|---|---|---|
-| _0 | uint256 | Delegator&#39;s unclaimed rewards with validator (in MATIC wei) |
-
-### openDelegatePosition
-
-```solidity
-function openDelegatePosition(address validator, uint256 durationWeeks) external payable
+function openVestedDelegatePosition(address validator, uint256 durationWeeks) external payable
 ```
 
 Delegates sent amount to validator. Set vesting position data. Delete old top-ups data if exists. Can be called by vesting positions&#39; managers only.
@@ -141,28 +94,6 @@ Delegates sent amount to validator. Add top-up data. Modify vesting position dat
 |---|---|---|
 | validator | address | Validator to delegate to |
 
-### totalDelegationOf
-
-```solidity
-function totalDelegationOf(address validator) external view returns (uint256)
-```
-
-Gets the total amount delegated to a validator.
-
-
-
-#### Parameters
-
-| Name | Type | Description |
-|---|---|---|
-| validator | address | Address of validator |
-
-#### Returns
-
-| Name | Type | Description |
-|---|---|---|
-| _0 | uint256 | Amount delegated (in MATIC wei) |
-
 ### undelegate
 
 ```solidity
@@ -187,7 +118,7 @@ Undelegates amount from validator for sender. Claims rewards beforehand.
 ### Delegated
 
 ```solidity
-event Delegated(address indexed delegator, address indexed validator, uint256 amount)
+event Delegated(address indexed validator, address indexed delegator, uint256 amount)
 ```
 
 
@@ -198,50 +129,14 @@ event Delegated(address indexed delegator, address indexed validator, uint256 am
 
 | Name | Type | Description |
 |---|---|---|
+| validator `indexed` | address | undefined |
 | delegator `indexed` | address | undefined |
-| validator `indexed` | address | undefined |
-| amount  | uint256 | undefined |
-
-### DelegatorRewardClaimed
-
-```solidity
-event DelegatorRewardClaimed(address indexed delegator, address indexed validator, bool indexed restake, uint256 amount)
-```
-
-
-
-
-
-#### Parameters
-
-| Name | Type | Description |
-|---|---|---|
-| delegator `indexed` | address | undefined |
-| validator `indexed` | address | undefined |
-| restake `indexed` | bool | undefined |
-| amount  | uint256 | undefined |
-
-### DelegatorRewardDistributed
-
-```solidity
-event DelegatorRewardDistributed(address indexed validator, uint256 amount)
-```
-
-
-
-
-
-#### Parameters
-
-| Name | Type | Description |
-|---|---|---|
-| validator `indexed` | address | undefined |
 | amount  | uint256 | undefined |
 
 ### Undelegated
 
 ```solidity
-event Undelegated(address indexed delegator, address indexed validator, uint256 amount)
+event Undelegated(address indexed validator, address indexed delegator, uint256 amount)
 ```
 
 
@@ -252,8 +147,8 @@ event Undelegated(address indexed delegator, address indexed validator, uint256 
 
 | Name | Type | Description |
 |---|---|---|
-| delegator `indexed` | address | undefined |
 | validator `indexed` | address | undefined |
+| delegator `indexed` | address | undefined |
 | amount  | uint256 | undefined |
