--- docs/elin/contracts/utils/Context.md
@@ -1,12 +1,0 @@
-# Context
-
-
-
-
-
-
-
-*Provides information about the current execution context, including the sender of the transaction and its data. While these are generally available via msg.sender and msg.data, they should not be accessed in such a direct manner, since when dealing with meta-transactions the account sending and paying for execution may not be the actual sender (as far as an application is concerned). This contract is only required for intermediate, library-like contracts.*
-
-
-
